+++ modules/project/import-ui/src/sap/commerce/toolset/project/HybrisProjectService.kt
@@ -1,6 +1,5 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
  * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
@@ -17,107 +16,97 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package sap.commerce.toolset.project;
+package sap.commerce.toolset.project
 
-import com.intellij.openapi.components.Service;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.idea.maven.model.MavenConstants;
-import sap.commerce.toolset.HybrisConstants;
-import sap.commerce.toolset.ccv2.CCv2Constants;
-import sap.commerce.toolset.project.descriptor.HybrisProjectDescriptor;
-import sap.commerce.toolset.project.vfs.VirtualFileSystemService;
-
-import java.io.File;
+import com.intellij.openapi.components.Service
+import com.intellij.openapi.components.service
+import com.intellij.util.application
+import org.jetbrains.idea.maven.model.MavenConstants
+import sap.commerce.toolset.HybrisConstants
+import sap.commerce.toolset.ccv2.CCv2Constants
+import sap.commerce.toolset.project.ProjectUtil.isHybrisModuleRoot
+import sap.commerce.toolset.project.descriptor.HybrisProjectDescriptor
+import sap.commerce.toolset.project.vfs.VirtualFileSystemService
+import java.io.File
 
 @Service
-public final class HybrisProjectService {
+class HybrisProjectService {
 
-    public boolean isConfigModule(@NotNull final File file) {
-        return new File(file, HybrisConstants.LOCAL_EXTENSIONS_XML).isFile();
-    }
+    fun isConfigModule(file: File) = File(file, "/licence/").isDirectory
+        && File(file, "/tomcat/tomcat_context.tpl").isFile
 
-    public boolean isCCv2Module(@NotNull final File file) {
-        return
-            (
-                file.getAbsolutePath().contains(CCv2Constants.CORE_CUSTOMIZE_NAME)
-                    || file.getAbsolutePath().contains(CCv2Constants.DATAHUB_NAME)
-                    || file.getAbsolutePath().contains(CCv2Constants.JS_STOREFRONT_NAME)
+    fun isCCv2Module(file: File): Boolean {
+        return (file.absolutePath.contains(CCv2Constants.CORE_CUSTOMIZE_NAME)
+            || file.absolutePath.contains(CCv2Constants.DATAHUB_NAME)
+            || file.absolutePath.contains(CCv2Constants.JS_STOREFRONT_NAME)
             )
-                && new File(file, CCv2Constants.MANIFEST_NAME).isFile();
+            && File(file, CCv2Constants.MANIFEST_NAME).isFile()
     }
 
-    public boolean isAngularModule(@NotNull final File file) {
-        return new File(file, HybrisConstants.FILE_ANGULAR_JSON).isFile();
-    }
+    fun isAngularModule(file: File) = File(file, HybrisConstants.FILE_ANGULAR_JSON).isFile()
 
-    public boolean isPlatformModule(@NotNull final File file) {
-        return file.getName().equals(HybrisConstants.EXTENSION_NAME_PLATFORM)
-            && new File(file, HybrisConstants.EXTENSIONS_XML).isFile();
-    }
+    fun isPlatformModule(file: File) = file.getName() == HybrisConstants.EXTENSION_NAME_PLATFORM
+        && File(file, HybrisConstants.EXTENSIONS_XML).isFile()
 
-    public boolean isPlatformExtModule(@NotNull final File file) {
-        return file.getAbsolutePath().contains(HybrisConstants.PLATFORM_EXT_MODULE_PREFIX)
-            && new File(file, HybrisConstants.EXTENSION_INFO_XML).isFile()
-            && !isCoreExtModule(file);
-    }
+    fun isPlatformExtModule(file: File) = file.absolutePath.contains(HybrisConstants.PLATFORM_EXT_MODULE_PREFIX)
+        && File(file, HybrisConstants.EXTENSION_INFO_XML).isFile()
+        && !isCoreExtModule(file)
 
-    public boolean isCoreExtModule(@NotNull final File file) {
-        return file.getAbsolutePath().contains(HybrisConstants.PLATFORM_EXT_MODULE_PREFIX)
-            && file.getName().equals(HybrisConstants.EXTENSION_NAME_CORE)
-            && new File(file, HybrisConstants.EXTENSION_INFO_XML).isFile();
-    }
+    fun isCoreExtModule(file: File) = file.absolutePath.contains(HybrisConstants.PLATFORM_EXT_MODULE_PREFIX)
+        && file.getName() == HybrisConstants.EXTENSION_NAME_CORE
+        && File(file, HybrisConstants.EXTENSION_INFO_XML).isFile()
 
-    public boolean isHybrisModule(@NotNull final File file) {
-        return ProjectUtil.isHybrisModuleRoot(file);
-    }
+    fun isHybrisModule(file: File): Boolean = isHybrisModuleRoot(file)
 
-    public boolean isOutOfTheBoxModule(@NotNull final File file, final HybrisProjectDescriptor rootProjectDescriptor) {
-        final File extDir = rootProjectDescriptor.getExternalExtensionsDirectory();
+    fun isOutOfTheBoxModule(file: File, rootProjectDescriptor: HybrisProjectDescriptor): Boolean {
+        val extDir = rootProjectDescriptor.externalExtensionsDirectory
         if (extDir != null) {
             if (VirtualFileSystemService.getInstance().fileContainsAnother(extDir, file)) {
                 // this will override bin/ext-* naming convention.
-                return false;
+                return false
             }
         }
-        return (file.getAbsolutePath().contains(HybrisConstants.PLATFORM_OOTB_MODULE_PREFIX) ||
-            file.getAbsolutePath().contains(HybrisConstants.PLATFORM_OOTB_MODULE_PREFIX_2019)
-        )
-            && new File(file, HybrisConstants.EXTENSION_INFO_XML).isFile();
+        return (file.absolutePath.contains(HybrisConstants.PLATFORM_OOTB_MODULE_PREFIX) ||
+            file.absolutePath.contains(HybrisConstants.PLATFORM_OOTB_MODULE_PREFIX_2019)
+            )
+            && File(file, HybrisConstants.EXTENSION_INFO_XML).isFile()
     }
 
-    public boolean isMavenModule(final File rootProjectDirectory) {
-        if (rootProjectDirectory.getAbsolutePath().contains(HybrisConstants.PLATFORM_MODULE_PREFIX)) {
-            return false;
+    fun isMavenModule(rootProjectDirectory: File): Boolean {
+        if (rootProjectDirectory.absolutePath.contains(HybrisConstants.PLATFORM_MODULE_PREFIX)) {
+            return false
         }
-        return new File(rootProjectDirectory, MavenConstants.POM_XML).isFile();
+        return File(rootProjectDirectory, MavenConstants.POM_XML).isFile()
     }
 
-    public boolean isEclipseModule(final File rootProjectDirectory) {
-        if (rootProjectDirectory.getAbsolutePath().contains(HybrisConstants.PLATFORM_MODULE_PREFIX)) {
-            return false;
+    fun isEclipseModule(rootProjectDirectory: File): Boolean {
+        if (rootProjectDirectory.absolutePath.contains(HybrisConstants.PLATFORM_MODULE_PREFIX)) {
+            return false
         }
-        return new File(rootProjectDirectory, HybrisConstants.DOT_PROJECT).isFile();
+        return File(rootProjectDirectory, HybrisConstants.DOT_PROJECT).isFile()
     }
 
-    public boolean isGradleModule(final File file) {
-        if (file.getAbsolutePath().contains(HybrisConstants.PLATFORM_MODULE_PREFIX)) {
-            return false;
+    fun isGradleModule(file: File): Boolean {
+        if (file.absolutePath.contains(HybrisConstants.PLATFORM_MODULE_PREFIX)) {
+            return false
         }
-        return new File(file, HybrisConstants.GRADLE_SETTINGS).isFile()
-            || new File(file, HybrisConstants.GRADLE_BUILD).isFile();
+        return File(file, HybrisConstants.GRADLE_SETTINGS).isFile()
+            || File(file, HybrisConstants.GRADLE_BUILD).isFile()
     }
 
-    public boolean isGradleKtsModule(final File file) {
-        if (file.getAbsolutePath().contains(HybrisConstants.PLATFORM_MODULE_PREFIX)) {
-            return false;
+    fun isGradleKtsModule(file: File): Boolean {
+        if (file.absolutePath.contains(HybrisConstants.PLATFORM_MODULE_PREFIX)) {
+            return false
         }
-        return new File(file, HybrisConstants.GRADLE_SETTINGS_KTS).isFile()
-            || new File(file, HybrisConstants.GRADLE_BUILD_KTS).isFile();
+        return File(file, HybrisConstants.GRADLE_SETTINGS_KTS).isFile()
+            || File(file, HybrisConstants.GRADLE_BUILD_KTS).isFile()
     }
 
-    public boolean hasVCS(final File rootProjectDirectory) {
-        return new File(rootProjectDirectory, ".git").isDirectory()
-            || new File(rootProjectDirectory, ".svn").isDirectory()
-            || new File(rootProjectDirectory, ".hg").isDirectory();
+    fun hasVCS(rootProjectDirectory: File?) = File(rootProjectDirectory, ".git").isDirectory()
+        || File(rootProjectDirectory, ".svn").isDirectory()
+        || File(rootProjectDirectory, ".hg").isDirectory()
+
+    companion object {
+        fun getInstance(): HybrisProjectService = application.service()
     }
 }